ci: use the correct SHA for the cz check
Was previously using the wrong SHA value in the `cz` or commitizen
check. This was due to the fact that when commits get merged by GitHub
they modify the commit on merge and thus change the SHA :(

This worked for awhile but eventually GitHub pruned the orphan commit
and then the CI started breaking.

This happened in https://github.com/LibreBooking/app/pull/583