fix(reservation): pdf generation is not working in French
When a reminder is set and the language is set to French, the reservation pdf
can't be generated because the template does not escape for javascript after
translation. In French the default string contains an apostrophe and this is
breaking the javascript code. It also makes the update of the reservation
impossible since the javascript code generate an error. Adding
"|escape:'javascript'" to ReminderBeforeStart and ReminderBeforeEnd
fix the problem. The code also fix the French translation.

Closes: #584