fix: use default logging level of 'error'
If no logging level is set then default to 'error'

Previously it was 'none'. In that case critical errors could be
happening but none of them would show up in the log.

We currently only support logging levels of 'none', 'debug', and
'error'. Remove unsupported levels.

Change logging levels to lowercase to fix bug as uppercase is incorrect.